make a new module auto_flush
