+++ src/lib.rs
@@ -137,6 +137,7 @@ extern crate coarsetime;
 #[macro_use]
 mod macros;
 mod atomic64;
+mod auto_flush;
 mod counter;
 mod desc;
 mod encoder;
@@ -159,15 +160,13 @@ pub mod local {
     Unsync local metrics, provides better performance.
 
     */
-    pub use super::counter::{
-        AFLDelegator, AFLocalCounter, LocalCounter, LocalCounterVec, LocalIntCounter,
-        LocalIntCounterVec,
-    };
-    pub use super::histogram::{
-        AFLHistogramDelegator, AFLocalHistogram, LocalHistogram, LocalHistogramTimer,
-        LocalHistogramVec,
-    };
+    pub use super::counter::{LocalCounter, LocalCounterVec, LocalIntCounter, LocalIntCounterVec};
+    pub use super::histogram::{LocalHistogram, LocalHistogramTimer, LocalHistogramVec};
     pub use super::metrics::{LocalMetric, MayFlush};
+
+    pub use super::auto_flush::{
+        AFLocalCounter, AFLocalHistogram, CounterDelegator, HistogramDelegator,
+    };
 }
 
 pub mod core {
